{blockchain/events,tokens/service}: revert chain types back to int64
The previous commit to update the block chain types to uint64 (as they
are never signed) conflicts with the satellite use of the service
endpoint API. Therefore the type change must be reverted so that the
service endpoint matches the satellite client.

+++ blockchain/events/config.go
@@ -5,8 +5,8 @@ package events
 
 // Config is a configuration struct for the transfer events service.
 type Config struct {
-	AddressBatchSize int    `help:"number of Addresses to fetch new events for in a single request" default:"100"`
-	BlockBatchSize   int    `help:"number of blocks to fetch new events for in a single request" default:"5000"`
-	ChainReorgBuffer uint64 `help:"minimum number of blocks to re-query for when looking for new transfer events" default:"15"`
-	MaximumQuerySize uint64 `help:"maximum number of blocks prior to the latest block that storjscan can query for" default:"10000"`
+	AddressBatchSize int `help:"number of Addresses to fetch new events for in a single request" default:"100"`
+	BlockBatchSize   int `help:"number of blocks to fetch new events for in a single request" default:"5000"`
+	ChainReorgBuffer int `help:"minimum number of blocks to re-query for when looking for new transfer events" default:"15"`
+	MaximumQuerySize int `help:"maximum number of blocks prior to the latest block that storjscan can query for" default:"10000"`
 }
